requirements.txt
psutil==5.8.0
pyzmq==21.0.1